# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personnel Monitoring System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export type LoginResponseRole =
  (typeof LoginResponseRole)[keyof typeof LoginResponseRole];

export const LoginResponseRole = {
  admin: "admin",
  user: "user",
} as const;

export interface LoginResponse {
  id: number;
  username: string;
  role: LoginResponseRole;
  personnelId?: number | null;
  name?: string | null;
}

export interface Personnel {
  id: number;
  lastName: string;
  firstName: string;
  middleInitial?: string | null;
  employeeId: string;
  department: string;
  position: string;
  photoUrl?: string | null;
  vehiclePlate?: string | null;
  hasAccount: boolean;
  createdAt: string;
}

export interface CreatePersonnelRequest {
  lastName: string;
  firstName: string;
  middleInitial?: string | null;
  employeeId: string;
  department: string;
  position: string;
  photoUrl?: string | null;
  vehiclePlate?: string | null;
  createAccount?: boolean;
  password?: string | null;
}

export interface UpdatePersonnelRequest {
  lastName?: string;
  firstName?: string;
  middleInitial?: string | null;
  department?: string;
  position?: string;
  photoUrl?: string | null;
  vehiclePlate?: string | null;
}

export type AccountRole = (typeof AccountRole)[keyof typeof AccountRole];

export const AccountRole = {
  admin: "admin",
  user: "user",
} as const;

export interface Account {
  id: number;
  username: string;
  role: AccountRole;
  personnelId?: number | null;
  personnelName?: string | null;
  createdAt: string;
}

export interface CreateAccountRequest {
  personnelId: number;
  password: string;
}

export interface UpdateAccountRequest {
  password: string;
}