# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Personnel Monitoring System API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: personnel
    description: Personnel management
  - name: accounts
    description: Account management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user session
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /personnel:
    get:
      operationId: listPersonnel
      tags: [personnel]
      summary: List all personnel
      responses:
        "200":
          description: List of personnel
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Personnel"
    post:
      operationId: createPersonnel
      tags: [personnel]
      summary: Create personnel and optionally an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePersonnelRequest"
      responses:
        "201":
          description: Personnel created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Personnel"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /personnel/{id}:
    get:
      operationId: getPersonnel
      tags: [personnel]
      summary: Get a personnel by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Personnel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Personnel"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updatePersonnel
      tags: [personnel]
      summary: Update personnel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePersonnelRequest"
      responses:
        "200":
          description: Updated personnel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Personnel"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deletePersonnel
      tags: [personnel]
      summary: Delete personnel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List all accounts
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create account for existing personnel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountRequest"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          description: Validation error or duplicate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /accounts/{id}:
    put:
      operationId: updateAccount
      tags: [accounts]
      summary: Update account password
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountRequest"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteAccount
      tags: [accounts]
      summary: Delete account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password

    LoginResponse:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [admin, user]
        personnelId:
          type: integer
          nullable: true
        name:
          type: string
          nullable: true
      required:
        - id
        - username
        - role

    Personnel:
      type: object
      properties:
        id:
          type: integer
        lastName:
          type: string
        firstName:
          type: string
        middleInitial:
          type: string
          nullable: true
        employeeId:
          type: string
        department:
          type: string
        position:
          type: string
        photoUrl:
          type: string
          nullable: true
        vehiclePlate:
          type: string
          nullable: true
        hasAccount:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - lastName
        - firstName
        - employeeId
        - department
        - position
        - hasAccount
        - createdAt

    CreatePersonnelRequest:
      type: object
      properties:
        lastName:
          type: string
        firstName:
          type: string
        middleInitial:
          type: string
          nullable: true
        employeeId:
          type: string
        department:
          type: string
        position:
          type: string
        photoUrl:
          type: string
          nullable: true
        vehiclePlate:
          type: string
          nullable: true
        createAccount:
          type: boolean
        password:
          type: string
          nullable: true
      required:
        - lastName
        - firstName
        - employeeId
        - department
        - position

    UpdatePersonnelRequest:
      type: object
      properties:
        lastName:
          type: string
        firstName:
          type: string
        middleInitial:
          type: string
          nullable: true
        department:
          type: string
        position:
          type: string
        photoUrl:
          type: string
          nullable: true
        vehiclePlate:
          type: string
          nullable: true

    Account:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [admin, user]
        personnelId:
          type: integer
          nullable: true
        personnelName:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - username
        - role
        - createdAt

    CreateAccountRequest:
      type: object
      properties:
        personnelId:
          type: integer
        password:
          type: string
      required:
        - personnelId
        - password

    UpdateAccountRequest:
      type: object
      properties:
        password:
          type: string
      required:
        - password