# lib/api-zod/src/generated/types/createPersonnelRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personnel Monitoring System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePersonnelRequest {
  lastName: string;
  firstName: string;
  middleInitial?: string | null;
  employeeId: string;
  department: string;
  position: string;
  photoUrl?: string | null;
  vehiclePlate?: string | null;
  createAccount?: boolean;
  password?: string | null;
}